Testing HashMap
Did some more work in testing out the functionality of a HashMap in modeling the board. I'm having some trouble dealing with the abstraction's immutability, and working with a other elements of the `Board` struct which are determined by the `mines` element.

I'm not sure exactly what is going wrong.