Major overhaul of board.rs
Switched around the abstractions and data structures used to represent `Board`—no longer using HashMaps, but now a compilation-time-known constants in Prelude (`WIDTH`, and `HEIGHT`).

Also did a different structuring of the modules below it, so that the two accessible structs are `pub`, whereas the enums which are contained are `pub(super)`.